backend/requirements.txt
Flask
librosa
dtw-python
gradio_client
cassandra-driver[gevent]
numpy